NXP-27159: fix MongoDB/Hibernate audit query on dates
